Remove InitWaku component
There was an issue where the observers are added/removed continously.
This is due to using `useEffect` on props.

By removing this component then `useEffect` ends only being called when
 waku changes, ie, at initialisation.
