+++ examples/eth-dm/src/waku.ts
@@ -8,87 +8,7 @@ import { byteArrayToHex } from './utils';
 export const PublicKeyContentTopic = '/eth-dm/1/public-key/proto';
 export const DirectMessageContentTopic = '/eth-dm/1/direct-message/json';
 
-interface Props {
-  waku: Waku | undefined;
-  setWaku: (waku: Waku) => void;
-  ethDmKeyPair: KeyPair | undefined;
-  setPublicKeys: Dispatch<SetStateAction<Map<string, string>>>;
-  setMessages: Dispatch<SetStateAction<Message[]>>;
-  address: string | undefined;
-}
-
-/**
- * Does all the waku initialization
- */
-export default function InitWaku({
-  waku,
-  setWaku,
-  ethDmKeyPair,
-  setPublicKeys,
-  setMessages,
-  address,
-}: Props) {
-  useEffect(() => {
-    if (waku) return;
-    initWaku()
-      .then((wakuNode) => {
-        console.log('waku: ready');
-        setWaku(wakuNode);
-      })
-      .catch((e) => {
-        console.error('Failed to initiate Waku', e);
-      });
-  }, [waku, setWaku]);
-
-  const observerPublicKeyMessage = handlePublicKeyMessage.bind(
-    {},
-    ethDmKeyPair?.publicKey,
-    setPublicKeys
-  );
-
-  const observerDirectMessage =
-    ethDmKeyPair && address
-      ? handleDirectMessage.bind(
-          {},
-          setMessages,
-          ethDmKeyPair.privateKey,
-          address
-        )
-      : undefined;
-
-  useEffect(() => {
-    if (!waku) return;
-    waku.relay.addObserver(observerPublicKeyMessage, [PublicKeyContentTopic]);
-
-    return function cleanUp() {
-      if (!waku) return;
-      waku.relay.deleteObserver(observerPublicKeyMessage, [
-        PublicKeyContentTopic,
-      ]);
-    };
-  });
-
-  useEffect(() => {
-    if (!waku) return;
-    if (!observerDirectMessage) return;
-    waku.relay.addObserver(observerDirectMessage, [DirectMessageContentTopic]);
-
-    return function cleanUp() {
-      if (!waku) return;
-      if (!observerDirectMessage) return;
-      waku.relay.deleteObserver(observerDirectMessage, [
-        DirectMessageContentTopic,
-      ]);
-    };
-  });
-
-  // Returns an empty fragment.
-  // Taking advantages of React's state management and useEffect()
-  // Not sure it is best practice but it works.
-  return <></>;
-}
-
-async function initWaku(): Promise<Waku> {
+export async function initWaku(): Promise<Waku> {
   const waku = await Waku.create({});
 
   const nodes = await getNodes();
@@ -105,7 +25,7 @@ function getNodes() {
   return getStatusFleetNodes(Environment.Prod);
 }
 
-function handlePublicKeyMessage(
+export function handlePublicKeyMessage(
   myPublicKey: string | undefined,
   setter: Dispatch<SetStateAction<Map<string, string>>>,
   msg: WakuMessage
@@ -128,7 +48,7 @@ function handlePublicKeyMessage(
   }
 }
 
-async function handleDirectMessage(
+export async function handleDirectMessage(
   setter: Dispatch<SetStateAction<Message[]>>,
   privateKey: string,
   address: string,
